Bound numeric literals by Numeric, not their own class
The type 10 admits every value it is == to, and Ruby numeric equality
crosses classes: 10 == 10.0 == 10r == BigDecimal(10). So Integer was
never a sound bound for the literal 10 — a fuzz comparing subtype?
against value-level inclusion flagged exactly this. The tightest honest
nominal bound is Numeric. Class-anchored literals are spelled
_Integer(10), _SameObject(10) or _Constraint(10, integer?: true), all of
which are sound subtypes of Integer.

Serialization is unaffected in behavior but now dispatches const types
by canonical representation (serializer.type === type, including unions
of consts) rather than through the subtype lattice, since the wire
format of the value 1 is an integer even though the type 1 is not
bounded by Integer.

+++ test/types/_tuple.test.rb
@@ -16,7 +16,7 @@ test "hierarchy" do
 	assert_subtype _Tuple(String, Integer), _Tuple(String, Integer)
 	assert_subtype _Tuple(String, Integer), _Tuple(String, Numeric)
 	assert_subtype _Tuple(_String(length: 10), Integer), _Tuple(_String(length: 5..15), Numeric)
-	assert_subtype _Tuple(10, :a), _Tuple(Integer, Symbol)
+	assert_subtype _Tuple(10, :a), _Tuple(Numeric, Symbol)
 
 	refute_subtype _Tuple(String, Float), _Tuple(String, Integer)
 	refute_subtype _Tuple(String, Numeric), _Tuple(String, Integer)
